requirements.txt
# Dependencias principales
songbloom @ git+https://github.com/suno-ai/song-bloom.git
retrieval-based-voice-conversion @ git+https://github.com/RVC-Project/Retrieval-based-Voice-Conversion-WebUI.git
bark @ git+https://github.com/suno-ai/bark.git
matchering
lyricjam
omi @ git+https://github.com/BasedHardware/omi.git
chroma-hnswlib
fastapi
python-telegram-bot
requests
sqlalchemy
